+++ argtyped/arguments.py
@@ -1,8 +1,22 @@
 import argparse
 import functools
+import shutil
 import sys
+from abc import ABCMeta
 from collections import OrderedDict
-from typing import Any, Callable, Dict, List, Optional, TypeVar
+from typing import (
+    Any,
+    Callable,
+    Dict,
+    List,
+    Literal,
+    NamedTuple,
+    Optional,
+    Tuple,
+    Type,
+    TypeVar,
+    Union,
+)
 
 from .custom_types import (
     Switch,
@@ -15,6 +29,8 @@ from .custom_types import (
 
 __all__ = [
     "Arguments",
+    "ArgumentSpec",
+    "argument_specs",
 ]
 
 T = TypeVar("T")
@@ -64,11 +80,11 @@ class ArgumentParser(argparse.ArgumentParser):
 
 
 def _bool_conversion_fn(s: str) -> bool:
-    if s.lower() in ["y", "yes", "true", "ok"]:
+    if s.lower() in {"y", "yes", "true", "ok"}:
         return True
-    if s.lower() in ["n", "no", "false"]:
+    if s.lower() in {"n", "no", "false"}:
         return False
-    raise ValueError(f"Invalid value '{s}' for bool argument")
+    raise ValueError(f"Invalid value {s!r} for bool argument")
 
 
 def _optional_wrapper_fn(fn: ConversionFn[T]) -> ConversionFn[Optional[T]]:
@@ -86,8 +102,162 @@ _TYPE_CONVERSION_FN: Dict[type, ConversionFn[Any]] = {
 }
 
 
-class Arguments:
-    r"""A typed version of ``argparse``. It's easier to illustrate using an example:
+class ArgumentSpec(NamedTuple):
+    nullable: bool
+    required: bool
+    value_type: Union[type, ConversionFn[Any]]
+    type: Literal["normal", "switch"]
+    choices: Optional[Tuple[str, ...]] = None
+    default: Optional[Any] = None
+
+
+class ArgumentsMeta(ABCMeta):
+    r"""
+    Metaclass for :class:`Arguments`. The type annotations are parsed and converted into
+    an ``argparse.ArgumentParser`` on class creation.
+    """
+    __parser__: Optional[ArgumentParser]
+    __arguments__: "OrderedDict[str, ArgumentSpec]"
+
+    def __new__(mcs, name, bases, namespace, **kwargs):
+        cls: "Type[Arguments]" = super().__new__(mcs, name, bases, namespace)
+
+        root = kwargs.get("_root", False)
+        if not root and not issubclass(cls, Arguments):
+            raise TypeError(f"Type {cls.__name__!r} must inherit from `Arguments`")
+        if root:
+            cls.__parser__ = None
+            cls.__arguments__ = {}
+            return cls
+
+        arguments: "OrderedDict[str, ArgumentSpec]" = OrderedDict()
+        for base in reversed(bases):
+            # Use reversed order so derived classes can override base annotations.
+            if issubclass(base, Arguments):
+                arguments.update(argument_specs(base))
+
+        # Check if there are arguments with default values but without annotations.
+        annotations = getattr(cls, "__annotations__", {})
+        for key in annotations:
+            if key.startswith("_") and not key.startswith("__"):
+                raise TypeError(f"Argument {key!r} must not start with an underscore")
+        annotations = OrderedDict(
+            [(k, v) for k, v in annotations.items() if not k.startswith("__")]
+        )
+        for key, value in namespace.items():
+            if not key.startswith("_") and not callable(value):
+                if key not in annotations and key not in arguments:
+                    raise TypeError(f"Argument {key!r} does not have type annotation")
+
+        # Check validity of arguments and create specs.
+        for arg_name, arg_typ in annotations.items():
+            has_default = arg_name in namespace
+            default_val = namespace.get(arg_name, None)
+            nullable = is_optional(arg_typ)
+            if nullable:
+                # extract the type wrapped inside `Optional`
+                arg_typ = unwrap_optional(arg_typ)
+
+            required = False
+            if nullable and not has_default:
+                has_default = True
+                default_val = None
+            elif not nullable and not has_default:
+                required = True
+
+            if not nullable and has_default and default_val is None:
+                raise TypeError(
+                    f"Argument {arg_name!r} has default value of None, but is not "
+                    f"nullable. Change type annotation to `Optional[...]` to allow "
+                    f"values of `None`"
+                )
+
+            if arg_typ is Switch:  # type: ignore[misc]
+                if not isinstance(default_val, bool):
+                    raise TypeError(
+                        f"Switch argument {arg_name!r} must have a default value of "
+                        f"type bool"
+                    )
+                spec = ArgumentSpec(
+                    type="switch",
+                    nullable=False,
+                    required=False,
+                    value_type=bool,
+                    default=default_val,
+                )
+            else:
+                if is_enum(arg_typ):
+                    value_type = arg_typ
+                    choices = tuple(x.name for x in arg_typ)
+                    if has_default and not isinstance(default_val, arg_typ):
+                        raise TypeError(
+                            f"Invalid default value for argument {arg_name!r}"
+                        )
+                elif is_choices(arg_typ):
+                    value_type = str
+                    choices = unwrap_choices(arg_typ)
+                    if any(not isinstance(choice, str) for choice in choices):
+                        raise TypeError("All choices must be strings")
+                    if has_default and default_val not in choices:
+                        raise TypeError(
+                            f"Invalid default value for argument {arg_name!r}"
+                        )
+                else:
+                    if arg_typ not in _TYPE_CONVERSION_FN and not callable(arg_typ):
+                        raise TypeError(
+                            f"Invalid type {arg_typ!r} for argument {arg_name!r}"
+                        )
+                    value_type = arg_typ
+                    choices = None
+                spec = ArgumentSpec(
+                    type="normal",
+                    nullable=nullable,
+                    required=required,
+                    value_type=value_type,
+                    choices=choices,
+                    default=default_val,
+                )
+            arguments[arg_name] = spec
+
+        # The parser will be lazily constructed when the `Arguments` instance is first
+        # initialized.
+        cls.__parser__ = None
+        cls.__arguments__ = arguments
+        return cls
+
+    def build_parser(cls):
+        parser = ArgumentParser()
+        for name, spec in cls.__arguments__.items():
+            arg_name = "--" + name.replace("_", "-")
+            if spec.type == "normal":
+                arg_type = spec.value_type
+                conversion_fn = _TYPE_CONVERSION_FN.get(arg_type, arg_type)
+                if spec.nullable:
+                    conversion_fn = _optional_wrapper_fn(conversion_fn)
+                kwargs = {
+                    "required": spec.required,
+                    "type": conversion_fn,
+                }
+                if not spec.required:
+                    kwargs["default"] = spec.default
+                parser.add_argument(arg_name, **kwargs)
+            else:
+                assert spec.default is not None
+                parser.add_switch_argument(arg_name, spec.default)
+
+        if cls.__module__ != "__main__":
+            # Usually arguments are defined in the same script that is directly
+            # run (__main__). If this is not the case, add a note in help message
+            # indicating where the arguments are defined.
+            parser.epilog = (
+                f"Note: Arguments defined in {cls.__module__}.{cls.__name__}"
+            )
+        cls.__parser__ = parser
+
+
+class Arguments(metaclass=ArgumentsMeta, _root=True):
+    r"""
+    A typed version of ``argparse``. It's easier to illustrate using an example:
 
     .. code-block:: python
 
@@ -171,114 +341,28 @@ class Arguments:
     actions are not supported.
     """
 
-    _annotations: "OrderedDict[str, type]"
-
     def __init__(self, args: Optional[List[str]] = None):
-        annotations: "OrderedDict[str, type]" = OrderedDict()
-        for base in reversed(self.__class__.mro()):
-            # Use reversed order so derived classes can override base annotations.
-            if base not in [object, Arguments]:
-                annotations.update(base.__dict__.get("__annotations__", {}))
-
-        # Check if there are arguments with default values but without annotations.
-        for key in dir(self):
-            value = getattr(self, key)
-            if not key.startswith("__") and not callable(value):
-                if key not in annotations:
-                    raise ValueError(f"Argument '{key}' does not have type annotation")
-
-        parser = ArgumentParser()
-        for arg_name, arg_typ in annotations.items():
-            # Check validity of name and type.
-
-            has_default = hasattr(self.__class__, arg_name)
-            default_val = getattr(self.__class__, arg_name, None)
-            nullable = is_optional(arg_typ)
-            if nullable:
-                # extract the type wrapped inside `Optional`
-                arg_typ = unwrap_optional(arg_typ)
-
-            required = False
-            if nullable and not has_default:
-                has_default = True
-                default_val = None
-            elif not nullable and not has_default:
-                required = True
-
-            if not nullable and has_default and default_val is None:
-                raise ValueError(
-                    f"Argument '{arg_name}' has default value of None, but is not "
-                    f"nullable"
-                )
-
-            parser_arg_name = "--" + arg_name.replace("_", "-")
-            parser_kwargs: Dict[str, Any] = {
-                "required": required,
-            }
-            if arg_typ is Switch:  # type: ignore[misc]
-                if not isinstance(default_val, bool):
-                    raise ValueError(
-                        f"Switch argument '{arg_name}' must have a default value of "
-                        f"type bool"
-                    )
-                parser.add_switch_argument(parser_arg_name, default_val)
-            elif is_choices(arg_typ) or is_enum(arg_typ):
-                if is_enum(arg_typ):
-                    choices = list(arg_typ)  # type: ignore[call-overload]
-                    parser_kwargs["type"] = arg_typ
-                else:
-                    choices = unwrap_choices(arg_typ)
-                    if any(not isinstance(choice, str) for choice in choices):
-                        raise ValueError("All choices must be strings")
-
-                parser_kwargs["choices"] = choices
-                if has_default:
-                    if default_val not in choices:
-                        raise ValueError(
-                            f"Invalid default value for argument '{arg_name}'"
-                        )
-                    parser_kwargs["default"] = default_val
-                parser.add_argument(parser_arg_name, **parser_kwargs)
-            else:
-                if arg_typ not in _TYPE_CONVERSION_FN and not callable(arg_typ):
-                    raise ValueError(
-                        f"Invalid type '{arg_typ}' for argument '{arg_name}'"
-                    )
-                conversion_fn = _TYPE_CONVERSION_FN.get(arg_typ, arg_typ)
-                if nullable:
-                    conversion_fn = _optional_wrapper_fn(conversion_fn)
-                parser_kwargs["type"] = conversion_fn
-                if has_default:
-                    parser_kwargs["default"] = default_val
-                parser.add_argument(parser_arg_name, **parser_kwargs)
-
-        if self.__class__.__module__ != "__main__":
-            # Usually arguments are defined in the same script that is directly
-            # run (__main__). If this is not the case, add a note in help message
-            # indicating where the arguments are defined.
-            parser.epilog = (
-                f"Note: Arguments defined in "
-                f"{self.__class__.__module__}.{self.__class__.__name__}"
-            )
-
-        namespace = parser.parse_args(args)
-        self._annotations = annotations
-        for arg_name, arg_typ in annotations.items():
+        if self.__parser__ is None:
+            self.__class__.build_parser()
+        namespace = self.__parser__.parse_args(args)
+        for arg_name in self.__arguments__:
             setattr(self, arg_name, getattr(namespace, arg_name))
 
     def to_dict(self) -> "OrderedDict[str, Any]":
-        r"""Convert the set of arguments to a dictionary.
+        r"""
+        Convert the set of arguments to a dictionary.
 
         :return: An ``OrderedDict`` mapping argument names to values.
         """
         return OrderedDict(
-            [(key, getattr(self, key)) for key in self._annotations.keys()]
+            [(key, getattr(self, key)) for key in argument_specs(self.__class__).keys()]
         )
 
     def to_string(
         self, width: Optional[int] = None, max_width: Optional[int] = None
     ) -> str:
-        r"""Represent the arguments as a table.
+        r"""
+        Represent the arguments as a table.
 
         :param width: Width of the printed table. Defaults to ``None``, which fits the
             table to its contents. An exception is raised when the table cannot be drawn
@@ -291,7 +375,7 @@ class Arguments:
 
         k_col = "Arguments"
         v_col = "Values"
-        valid_keys = list(self._annotations.keys())
+        valid_keys = list(argument_specs(self.__class__).keys())
         valid_vals = [repr(getattr(self, k)) for k in valid_keys]
         max_key = max(len(k_col), max(len(k) for k in valid_keys))
         max_val = max(len(v_col), max(len(v) for v in valid_vals))
@@ -317,3 +401,15 @@ class Arguments:
             s += get_row(k, v)
         s += f"╚═{'═' * max_key}═╧═{'═' * max_val}═╝\n"
         return s
+
+    def __repr__(self) -> str:
+        columns, _ = shutil.get_terminal_size()
+        return self.to_string(max_width=columns)
+
+
+def argument_specs(args_class: Type[Arguments]) -> "OrderedDict[str, ArgumentSpec]":
+    r"""
+    Return a dictionary mapping argument names to their specs (:class:`ArgumentSpec`
+    objects).
+    """
+    return args_class.__arguments__