+++ argtyped/custom_types.py
@@ -1,5 +1,5 @@
 import enum
-from collections import Iterable as IterableType
+from collections.abc import Iterable as IterableType
 from typing import Any, Iterable, Optional, Tuple, Type, TypeVar, Union
 
 __all__ = [
@@ -91,11 +91,10 @@ if HAS_LITERAL:
         metaclass.
         """
         return (
-                isinstance(typ, _Choices)
-                or getattr(typ, "__origin__", None) is Literal
-                or type(typ) is _Literal  # pylint: disable=unidiomatic-typecheck
+            isinstance(typ, _Choices)
+            or getattr(typ, "__origin__", None) is Literal
+            or type(typ) is _Literal  # pylint: disable=unidiomatic-typecheck
         )
-
 
     def unwrap_choices(typ: type) -> Tuple[str, ...]:
         r"""
@@ -114,7 +113,6 @@ else:
         checked using traditional methods, since :class:`Choices` is a metaclass.
         """
         return isinstance(typ, _Choices)
-
 
     def unwrap_choices(typ: type) -> Tuple[str, ...]:
         r""" Return the string literals associated with the choices type. """