+++ tests/test_errors.py
@@ -8,34 +8,37 @@ from argtyped import *
 
 
 def test_no_type_annotation():
-    class Args(Arguments):
-        a = 1  # wrong
-        b = 2  # wrong
-        c: int = 3
+    with pytest.raises(TypeError, match=r"does not have type annotation"):
 
-    with pytest.raises(ValueError, match=r"does not have type annotation"):
+        class Args(Arguments):
+            a = 1  # wrong
+            b = 2  # wrong
+            c: int = 3
+
         _ = Args()
 
 
 def test_non_nullable():
-    class Args(Arguments):
-        a: Optional[int] = None
-        b: Optional[int]
-        c: int = None  # wrong
+    with pytest.raises(TypeError, match=r"not nullable"):
 
-    with pytest.raises(ValueError, match=r"not nullable"):
+        class Args(Arguments):
+            a: Optional[int] = None
+            b: Optional[int]
+            c: int = None  # wrong
+
         _ = Args()
 
 
 def test_switch_not_bool():
-    class Args(Arguments):
-        a: Switch = True
-        b: Switch  # wrong
-        c: Switch = 0  # wrong
-
     with pytest.raises(
-        ValueError, match=r"Switch argument .* default value of type bool"
+        TypeError, match=r"Switch argument .* default value of type bool"
     ):
+
+        class Args(Arguments):
+            a: Switch = True
+            b: Switch  # wrong
+            c: Switch = 0  # wrong
+
         _ = Args()
 
 
@@ -45,21 +48,21 @@ def test_invalid_choice():
         class Args1(Arguments):
             a: Choices[()]
 
-    with pytest.raises(ValueError, match=r"Invalid default value"):
+    with pytest.raises(TypeError, match=r"Invalid default value"):
 
         class Args2(Arguments):
             a: Choices["a"] = "b"  # wrong
 
         _ = Args2()
 
-    with pytest.raises(ValueError, match=r"must be string"):
+    with pytest.raises(TypeError, match=r"must be string"):
 
         class Args3(Arguments):
             a: Choices["1", 2]
 
         _ = Args3()
 
-    with pytest.raises(ValueError, match=r"must be string"):
+    with pytest.raises(TypeError, match=r"must be string"):
 
         class Args4(Arguments):
             a: Literal["1", 2]
@@ -81,14 +84,14 @@ def test_invalid_bool(capsys):
 
 
 def test_invalid_type():
-    with pytest.raises(ValueError, match="Invalid type"):
+    with pytest.raises(TypeError, match="Invalid type"):
 
         class Args(Arguments):
             a: 5 = 0
 
         _ = Args()
 
-    with pytest.raises(ValueError, match="Invalid type"):
+    with pytest.raises(TypeError, match="Invalid type"):
 
         class Args(Arguments):
             b: "str" = 1
@@ -101,8 +104,9 @@ def test_invalid_enum():
         a = auto()
         b = auto()
 
-    class Args(Arguments):
-        enum: MyEnum = "c"
+    with pytest.raises(TypeError, match="Invalid default value"):
 
-    with pytest.raises(ValueError, match="Invalid default value"):
+        class Args(Arguments):
+            enum: MyEnum = "c"
+
         _ = Args()